+++ bfabric/entities/__init__.py
@@ -1,4 +1,6 @@
 from .dataset import Dataset
+from .parameter import Parameter
+from .workunit import Workunit
 
 
-__all__ = ["Dataset"]
+__all__ = ["Dataset", "Parameter", "Workunit"]