+++ bfabric/scripts/bfabric_list_not_available_proteomics_workunits.py
@@ -59,11 +59,7 @@ def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], config: 
 
 def list_not_available_proteomics_workunits(date_cutoff: datetime) -> None:
     """Lists proteomics work units that are not available on bfabric."""
-    client = Bfabric.from_config(verbose=True)
-    Console(stderr=True).print(
-        f"--- list not available proteomics work units created after {date_cutoff}---",
-        style="bright_yellow",
-    )
+    client = Bfabric.from_config()
 
     workunits_by_status = {}
     for status in ["Pending", "Processing", "Failed"]: