+++ bfabric/scripts/bfabric_save_fasta.py
@@ -17,7 +17,7 @@ BFABRIC_APPLICATION_ID = 61
 
 def save_fasta(container_id: int, fasta_file: Path) -> None:
     """Save a fasta file to bfabric."""
-    client = Bfabric.from_config(verbose=True)
+    client = Bfabric.from_config()
 
     print("Reading description from stdin")
     description = sys.stdin.read()