+++ bfabric/scripts/bfabric_setWorkunitStatus.py
@@ -20,7 +20,7 @@ def main_generic(result_status: str) -> None:
     parser = argparse.ArgumentParser(description=f"Sets workunit status to '{result_status}'")
     parser.add_argument("workunit_id", type=int, help="workunit id")
     args = parser.parse_args()
-    client = Bfabric.from_config(verbose=True)
+    client = Bfabric.from_config()
     res = client.save("workunit", {"id": args.workunit_id, "status": "available"})
     print(json.dumps(res.to_list_dict(), indent=2))