+++ docs/changelog.md
@@ -10,6 +10,10 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[tba\] - tba
 
+### Added
+
+- Add `Workunit` and `Parameter` entities.
+
 ## \[1.13.3\] - 2024-07-18
 
 ### Added